Move get_pivot_role_name and get_external_id to OrganizationService, as no environment related stuff inside
